requirements.txt
PyYAML
requests==2.22.0
Django==2.2.4
pytz==2019.2
djangorestframework==3.10.2
